requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-filter==24.3
django-import-export==4.3.2
django-simple-history==3.7.0
django-redis==5.4.0
redis==5.2.1
Pillow==11.0.0
psycopg2-binary==2.9.10
celery==5.3.6
drf-spectacular==0.28.0
django-environ==0.11.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
tablib==3.7.0
colorlog==6.7.0
termcolor==2.3.0
tablib[xls]
tablib[xlsx]
django-debug-toolbar==4.3.0
reportlab==4.2.5